// src/services/game-data-v2/attacks/types.ts
import type { OriginType, Sequence } from '@/services/game-data/types';

import type { DamageInstance } from '../damage-instances/types';

/**
 * A sequence of attacks that typically occurs in combat
 * Can be system/service created, or generated by the user
 */
export type Attack = {
  id: string;
  origin: 'system' | 'user';
  /** The user who created it, only populated when origin === 'user' */
  userId?: number;
  instances: Array<
    DamageInstance & {
      hitCount: number;
      alternativeDefinitions?: Partial<Record<Sequence, DamageInstance>>;
    }
  >;
};

export type EntityAttack = Attack & {
  entityId: number;
  skillId: number;
  skillOriginType: OriginType;
  skillName: string;
  skillDescription?: string;
  skillIconUrl?: string;
  skillAttribute?: {
    name: string;
    order: number;
  };
};